requirements.txt
streamlit
openai>=1.0.0
python-docx
requests
beautifulsoup4
python-dotenv
